Import from Apple Music (#25)
Importing a tracklist from Apple Music is as easy as pasting the album
URL into a new "Import From..." model. I have started with Apple
Music, but I will add more soon.

The Tracklist Editor still works entirely offline if you don't use the
import feature. However, due to CORS the import feature cannot be part
of the static web page. I will have to manually push changes to the
importer. If that becomes a pain I can automate it with CD later.

One known issue: If the album is semi-various artists, the importer will
only catch a few tracks. Semi-various artists is a thing with
Apple Music where albums that contain mostly one artist do not show that
artist on those tracks. This causes havoc with the (probably brittle)
regex that parses the raw HTML.